# docs/openapi.yaml
openapi: 3.0.3
info:
  title: YAWL REST API
  description: |
    RESTful API for YAWL (Yet Another Workflow Language) v5.2 Workflow Engine

    This API provides comprehensive workflow management capabilities including:
    - Session management and authentication
    - Work item lifecycle operations (checkout, checkin, complete)
    - Case management and control
    - Specification management and validation
    - Event subscriptions and monitoring
    - Extended operations (exception handling, task suspension/resumption)

    All operations require a valid session handle obtained through the `/ib/connect` endpoint.
  version: 5.2.0
  contact:
    name: YAWL Foundation
    url: https://www.yawlfoundation.org
  license:
    name: GNU Lesser General Public License v2.1
    url: https://www.gnu.org/licenses/old-licenses/lgpl-2.1.html
servers:
  - url: http://localhost:8080/yawl/api
    description: Local development server
    variables:
      port:
        default: '8080'
  - url: https://yawl.example.com/api
    description: Production server
...
paths:
...
  /ia/specifications:
    post:
      tags:
        - Specifications
      summary: Upload a workflow specification
      description: |
        **Note: Currently not implemented as REST endpoint.**
        Use the InterfaceA_EngineBasedServer servlet instead.

        Uploads and registers a new YAWL workflow specification.
      operationId: uploadSpecification
      deprecated: true
      parameters:
        - name: sessionHandle
          in: query
          required: true
          description: Valid session handle
          schema:
            type: string
      requestBody:
        required: true
        description: Workflow specification in YAWL XML format
        content:
          application/xml:
            schema:
              type: string
      responses:
        '501':
          description: Not implemented - use InterfaceA_EngineBasedServer servlet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Specifications
      summary: Get all loaded specifications
      description: |
        **Note: Currently not implemented as REST endpoint.**
        Use the InterfaceA_EngineBasedServer servlet instead.

        Retrieves metadata for all currently loaded specifications.
      operationId: getSpecifications
      deprecated: true
      parameters:
        - name: sessionHandle
          in: query
          required: true
          description: Valid session handle
          schema:
            type: string
      responses:
        '501':
          description: Not implemented - use InterfaceA_EngineBasedServer servlet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ia/specifications/{specId}:
    delete:
      tags:
        - Specifications
      summary: Unload a specification
      description: |
        **Note: Currently not implemented as REST endpoint.**
        Use the InterfaceA_EngineBasedServer servlet instead.

        Unloads a specification from the engine.
      operationId: unloadSpecification
      deprecated: true
      parameters:
        - name: specId
          in: path
          required: true
          description: Specification ID
          schema:
            type: string
        - name: sessionHandle
          in: query
          required: true
          description: Valid session handle
          schema:
            type: string
      responses:
        '501':
          description: Not implemented - use InterfaceA_EngineBasedServer servlet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing the failure
...
  securitySchemes:
    SessionHandle:
      type: apiKey
      in: query
      name: sessionHandle
      description: Session handle obtained from /ib/connect endpoint
security:
  - SessionHandle: []
...